I don't think I loved the edits, so using completion for now
