+++ server/src/data/proposal_templates/instructions.md
@@ -1,0 +1,19 @@
+Background:
+
+- My name is Brandi Alexander
+- I have my Bachelors Degree in Phsycology
+- Recently attneded a UX and UI Design Bootcamp
+- Agency name is Lavender Iquanas
+  - Started the Agency with my husband who is a developer
+
+Job Posting:
+
+{job_posting}
+
+Sample Proposal:
+
+{sample_proposal}
+
+----
+
+Task: Please edit the proposal to match the tone and style of the job posting. It is important that I sound like a human and not an AI. If the job listing asks specific questions be sure to answer them. If the job listing is specific about the type of business include details about that i the proposal
